labs/module-01/demos/demo-02-basic-streaming/pyproject.toml
[project]
name = "basic-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.2.0",
    "python-dotenv>=1.1.1",
]
